Update populate db
